Relocate server motd to MotdListener & cleanup plugin initialization
