+++ velocity/src/main/java/net/crystalixs/core/velocity/CorePlugin.java
@@ -2,22 +2,18 @@ package net.crystalixs.core.velocity;
 
 import com.velocitypowered.api.event.Subscribe;
 import com.velocitypowered.api.event.proxy.ProxyInitializeEvent;
-import com.velocitypowered.api.event.proxy.ProxyPingEvent;
 import com.velocitypowered.api.plugin.annotation.DataDirectory;
 import com.velocitypowered.api.proxy.ProxyServer;
-import com.velocitypowered.api.proxy.server.ServerPing;
 import jakarta.inject.Inject;
 import net.crystalixs.core.common.config.ObjectMapperFactory;
-import net.crystalixs.core.velocity.config.Motd;
 import net.crystalixs.core.velocity.config.VelocityConfig;
 import net.crystalixs.core.velocity.config.VelocityConfigLoader;
 import net.crystalixs.core.velocity.config.jackson.JacksonVelocity;
+import net.crystalixs.core.velocity.listener.MotdListener;
 import tools.jackson.databind.ObjectMapper;
 
 import java.nio.file.Path;
 import java.util.logging.Logger;
-
-import static net.kyori.adventure.text.Component.newline;
 
 public class CorePlugin {
 
@@ -35,29 +31,23 @@ public class CorePlugin {
 
     @Subscribe
     public void onProxyInitialization(ProxyInitializeEvent event) {
-        ObjectMapper mapper = ObjectMapperFactory.create(builder ->
-                builder.addModule(JacksonVelocity.builder()
-                        .withMiniMessage()
-                        .build()));
+        createOrLoadConfig();
+        registerListener(server);
+
+        logger.info("Velocity core plugin has been enabled!");
+    }
+
+    private void registerListener(ProxyServer server) {
+        server.getEventManager().register(this, new MotdListener(config));
+    }
+
+    private void createOrLoadConfig() {
+        JacksonVelocity jacksonVelocity = JacksonVelocity.builder().withMiniMessage().build();
+        ObjectMapper mapper = ObjectMapperFactory.create(builder -> builder.addModule(jacksonVelocity));
         VelocityConfigLoader loader = new VelocityConfigLoader(mapper, logger, dataDirectory.resolve("config.json"));
         loader.reload();
 
         config = loader.get();
-
-        server.getEventManager().register(this, this);
-        logger.info("Velocity core plugin has been enabled!");
     }
-
-    @Subscribe
-    public void onProxyPing(ProxyPingEvent event) {
-        Motd motd = config.motd();
-
-        ServerPing ping = event.getPing().asBuilder()
-                .description(motd.firstLine().append(newline()).append(motd.secondLine()))
-                .build();
-
-        event.setPing(ping);
-    }
-
 }
